// backend/src/types/amadeus.ts
export interface LocationSearchParams {
  keyword: string;
  subType?: string;
  countryCode?: string;
  view?: string;
  sort?: string;
  page?: {
    limit?: number;
    offset?: number;
  };
}

export interface Location {
  type: string;
  subType: string;
  name: string;
  detailedName: string;
  id: string;
  self: Self;
  timeZoneOffset: string;
  iataCode: string;
  geoCode: GeoCode;
  address: Address;
  analytics: Analytics;
}

export interface Self {
  href: string;
  methods: string[];
}

export interface GeoCode {
  latitude: number;
  longitude: number;
}

export interface Address {
  cityName: string;
  cityCode: string;
  countryName: string;
  countryCode: string;
  regionCode: string;
}

export interface Analytics {
  travelers: Travelers;
}

export interface Travelers {
  score: number;
}